feat(edgeless): edgeless copilot toolbar entry
