+++ packages/blocks/src/_common/theme/css-variables.ts
@@ -139,3 +139,7 @@ export type CssVariablesMap = Record<CssVariableName, string>;
 export function isCssVariable(name: string): name is CssVariableName {
   return StyleVariables.includes(name as CssVariableName);
 }
+
+export function isTransparent(color: CssVariableName) {
+  return color.toLowerCase() === '--affine-palette-transparent';
+}
