+++ packages/blocks/src/root-block/types.ts
@@ -6,6 +6,7 @@ import type { AFFINE_DOC_REMOTE_SELECTION_WIDGET } from './widgets/doc-remote-se
 import type { AFFINE_DRAG_HANDLE_WIDGET } from './widgets/drag-handle/drag-handle.js';
 import type { AFFINE_EDGELESS_REMOTE_SELECTION_WIDGET } from './widgets/edgeless-remote-selection/index.js';
 import type { AFFINE_EDGELESS_ZOOM_TOOLBAR_WIDGET } from './widgets/edgeless-zoom-toolbar/index.js';
+import type { EDGELESS_ELEMENT_TOOLBAR_WIDGET } from './widgets/element-toolbar/index.js';
 import type { AFFINE_FORMAT_BAR_WIDGET } from './widgets/format-bar/format-bar.js';
 import type { AFFINE_INNER_MODAL_WIDGET } from './widgets/inner-modal/inner-modal.js';
 import type { AFFINE_LINKED_DOC_WIDGET } from './widgets/linked-doc/index.js';
@@ -37,7 +38,8 @@ export type EdgelessRootBlockWidgetName =
   | typeof AFFINE_FORMAT_BAR_WIDGET
   | typeof AFFINE_DOC_REMOTE_SELECTION_WIDGET
   | typeof AFFINE_EDGELESS_REMOTE_SELECTION_WIDGET
-  | typeof AFFINE_EDGELESS_ZOOM_TOOLBAR_WIDGET;
+  | typeof AFFINE_EDGELESS_ZOOM_TOOLBAR_WIDGET
+  | typeof EDGELESS_ELEMENT_TOOLBAR_WIDGET;
 
 export type RootBlockComponent =
   | PageRootBlockComponent
