+++ packages/blocks/src/root-block/widgets/element-toolbar/change-group-button.ts
@@ -1,26 +1,26 @@
-import '../buttons/tool-icon-button.js';
+import '../../edgeless/components/buttons/tool-icon-button.js';
 import './component-toolbar-menu-divider.js';
 
 import { WithDisposable } from '@blocksuite/block-std';
 import { css, html, LitElement, nothing } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { toast } from '../../../../_common/components/toast.js';
+import { toast } from '../../../_common/components/toast.js';
 import {
   NoteIcon,
   RenameIcon,
   UngroupButtonIcon,
-} from '../../../../_common/icons/index.js';
-import { NoteDisplayMode } from '../../../../_common/types.js';
-import { matchFlavours } from '../../../../_common/utils/model.js';
-import type { GroupElementModel } from '../../../../surface-block/index.js';
+} from '../../../_common/icons/index.js';
+import { NoteDisplayMode } from '../../../_common/types.js';
+import { matchFlavours } from '../../../_common/utils/model.js';
+import type { GroupElementModel } from '../../../surface-block/index.js';
 import {
   deserializeXYWH,
   serializeXYWH,
-} from '../../../../surface-block/index.js';
-import type { SurfaceBlockComponent } from '../../../../surface-block/surface-block.js';
-import { DEFAULT_NOTE_HEIGHT } from '../../utils/consts.js';
-import { mountGroupTitleEditor } from '../../utils/text.js';
+} from '../../../surface-block/index.js';
+import type { EdgelessRootBlockComponent } from '../../edgeless/edgeless-root-block.js';
+import { DEFAULT_NOTE_HEIGHT } from '../../edgeless/utils/consts.js';
+import { mountGroupTitleEditor } from '../../edgeless/utils/text.js';
 
 @customElement('edgeless-change-group-button')
 export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
@@ -31,15 +31,15 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
   `;
 
   @property({ attribute: false })
-  surface!: SurfaceBlockComponent;
+  edgeless!: EdgelessRootBlockComponent;
 
   @property({ attribute: false })
   groups!: GroupElementModel[];
 
   private _insertIntoPage() {
-    if (!this.surface.doc.root) return;
+    if (!this.edgeless.doc.root) return;
 
-    const rootModel = this.surface.doc.root;
+    const rootModel = this.edgeless.doc.root;
     const notes = rootModel.children.filter(
       model =>
         matchFlavours(model, ['affine:note']) &&
@@ -56,7 +56,7 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
       targetXYWH[1] = targetXYWH[1] + targetXYWH[3];
       targetXYWH[3] = DEFAULT_NOTE_HEIGHT;
 
-      const newAddedNote = this.surface.doc.addBlock(
+      const newAddedNote = this.edgeless.doc.addBlock(
         'affine:note',
         {
           xywh: serializeXYWH(...targetXYWH),
@@ -67,7 +67,7 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
       targetParent = newAddedNote;
     }
 
-    this.surface.doc.addBlock(
+    this.edgeless.doc.addBlock(
       'affine:surface-ref',
       {
         reference: this.groups[0].id,
@@ -76,7 +76,7 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
       targetParent
     );
 
-    toast(this.surface.host, 'Group has been inserted into page');
+    toast(this.edgeless.host, 'Group has been inserted into page');
   }
 
   protected override render() {
@@ -94,8 +94,7 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
             <component-toolbar-menu-divider></component-toolbar-menu-divider>
             <edgeless-tool-icon-button
               class=${'edgeless-component-toolbar-group-rename-button'}
-              @click=${() =>
-                mountGroupTitleEditor(groups[0], this.surface.edgeless)}
+              @click=${() => mountGroupTitleEditor(groups[0], this.edgeless)}
               .tooltip=${'Rename'}
               .tipPosition=${'bottom'}
             >
@@ -109,7 +108,7 @@ export class EdgelessChangeGroupButton extends WithDisposable(LitElement) {
       <edgeless-tool-icon-button
         class=${'edgeless-component-toolbar-ungroup-button'}
         @click=${() => {
-          groups.forEach(group => this.surface.edgeless.service.ungroup(group));
+          groups.forEach(group => this.edgeless.service.ungroup(group));
         }}
         .tooltip=${'Ungroup'}
         .tipPosition=${'bottom'}
@@ -135,4 +134,17 @@ declare global {
   interface HTMLElementTagNameMap {
     'edgeless-change-group-button': EdgelessChangeGroupButton;
   }
+}
+
+export function renderGroupButton(
+  edgeless: EdgelessRootBlockComponent,
+  groups?: GroupElementModel[]
+) {
+  return groups?.length
+    ? html`<edgeless-change-group-button
+        .edgeless=${edgeless}
+        .groups=${groups}
+      >
+      </edgeless-change-group-button>`
+    : nothing;
 }
