+++ packages/blocks/src/root-block/widgets/element-toolbar/release-from-group-button.ts
@@ -1,20 +1,20 @@
-import '../buttons/tool-icon-button.js';
+import '../../edgeless/components/buttons/tool-icon-button.js';
 
 import { WithDisposable } from '@blocksuite/block-std';
 import { html, LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { ReleaseFromGroupButtonIcon } from '../../../../_common/icons/index.js';
-import { GroupElementModel } from '../../../../surface-block/element-model/group.js';
-import type { SurfaceBlockComponent } from '../../../../surface-block/surface-block.js';
+import { ReleaseFromGroupButtonIcon } from '../../../_common/icons/index.js';
+import { GroupElementModel } from '../../../surface-block/element-model/group.js';
+import type { EdgelessRootBlockComponent } from '../../edgeless/edgeless-root-block.js';
 
 @customElement('edgeless-release-from-group-button')
 export class EdgelessReleaseFromGroupButton extends WithDisposable(LitElement) {
   @property({ attribute: false })
-  surface!: SurfaceBlockComponent;
+  edgeless!: EdgelessRootBlockComponent;
 
   private _releaseFromGroup() {
-    const service = this.surface.edgeless.service;
+    const service = this.edgeless.service;
     const element = service.selection.firstElement;
 
     if (!(element.group instanceof GroupElementModel)) return;
@@ -51,4 +51,12 @@ declare global {
   interface HTMLElementTagNameMap {
     'edgeless-release-from-group-button': EdgelessReleaseFromGroupButton;
   }
+}
+
+export function renderReleaseFromGroupButton(
+  edgeless: EdgelessRootBlockComponent
+) {
+  return html`<edgeless-release-from-group-button
+    .edgeless=${edgeless}
+  ></edgeless-release-from-group-button>`;
 }
